+++ apps/chaingraph-frontend/src/components/flow/Flow.tsx
@@ -36,7 +36,6 @@ import { Background, ReactFlow, useReactFlow } from '@xyflow/react'
 import { useUnit } from 'effector-react'
 import { AnimatePresence } from 'framer-motion'
 import { memo, use, useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import { FPSCounter } from './components/FPSCounter'
 
 // Configuration constants
 const defaultViewport: Viewport = {
@@ -249,11 +248,10 @@ function Flow({
           <FlowControlPanel />
         )}
 
-        <div className="absolute top-4 left-4 z-50">
-          {/* <ExecutionComponent /> */}
-
-          <FPSCounter />
-        </div>
+        {/* <div className="absolute top-4 left-4 z-50"> */}
+        {/* <ExecutionComponent /> */}
+        {/* <FPSCounter /> */}
+        {/* </div> */}
       </ReactFlow>
 
       {/* Context Menu */}